+++ myApp/interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/UserService.java
@@ -46,7 +46,7 @@ public interface UserService {
      * @param mail The mail of the user.
      * @return The created user.
      */
-    Optional<User> create(String username, String mail, String password, String description, String schedule, int userole);
+    Optional<User> create(String username, String mail, String password, String description, String schedule, Long userole);
 
     Optional<User> findByEmail(String mail);
